v3: better task metadata errors (#991)
* v3: better handle task metadata parse errors, and display nicely formatted errors (dev, deploy, UI)

* Add changeset

+++ packages/cli-v3/src/workers/prod/backgroundWorker.ts
@@ -19,7 +19,7 @@ import {
 } from "@trigger.dev/core/v3";
 import { Evt } from "evt";
 import { ChildProcess, fork } from "node:child_process";
-import { UncaughtExceptionError } from "../common/errors";
+import { TaskMetadataParseError, UncaughtExceptionError } from "../common/errors";
 
 class UnexpectedExitError extends Error {
   constructor(public code: number) {
@@ -146,6 +146,14 @@ export class ProdBackgroundWorker {
               clearTimeout(timeout);
               resolved = true;
               reject(new UncaughtExceptionError(message.error, message.origin));
+              child.kill();
+            }
+          },
+          TASKS_FAILED_TO_PARSE: async (message) => {
+            if (!resolved) {
+              clearTimeout(timeout);
+              resolved = true;
+              reject(new TaskMetadataParseError(message.zodIssues, message.tasks));
               child.kill();
             }
           },
